Changes in the code creating some example objects to try and have a moving robot (NOT FINISHED)

+++ include/encoders/AbstractEncoder.h
@@ -11,20 +11,20 @@ public:
     /**
      * @return the current encoder count
      */
-    virtual int64_t getEncoderCount() const = 0;
+    virtual int32_t getEncoderCount() const = 0;
     /**
      * This method can be use to get the delta count from the last call of this function
      * Its behavior should be that if the setEncoderCount is used then the delta should continue to work
      * @return the number of count since last call
      */
-    virtual int64_t getDeltaCount() = 0;
+    virtual int32_t getDeltaCount() = 0;
 
     /**
      * This method allows to set the count
      * @param count the count we want to set
      * @return the previous count
      */
-    virtual int64_t setEncoderCount(int64_t count) = 0;
+    virtual int32_t setEncoderCount(int32_t count) = 0;
 };
